+++ .gitignore
@@ -1,2 +1,4 @@
 /venv/
 /.pytest_cache/
+/.idea/misc.xml
+/.idea/other.xml
