+++ experiments/parameter_exploration/adaptation_exploration.py
@@ -2,13 +2,15 @@ from build.config import SUBSTRATE_TYPE, GAP_INV, CUSTOM_FIRST, CUSTOM_SECOND, R
     STEP_AMOUNT, X_STEP_POSSIBILITY, Y_STEP_POSSIBILITY, SIGMA, FORCE, ADAPTATION_ENABLED, ADAPTATION_MU, \
     ADAPTATION_LAMBDA, ADAPTATION_HISTORY, SIGMOID_STEEPNESS, FORWARD_SIG, REVERSE_SIG, FF_INTER, FT_INTER, SIGMOID_SHIFT
 from build import object_factory
-import visualization.visualization as vz
+import visualization as vz
 
 
 def adaptation_comparison_lambda():
     gc = []
 
     simulation = object_factory.build_simulation(GAP_INVERTED_LAMBDA_1)
+
+    vz.visualize_substrate(simulation.substrate)
     simulation.run()
     gc.extend(simulation.growth_cones)
