+++ experiments/parameter_exploration/results/history_comparison.png
