+++ experiments/parameter_exploration/results/lamba_comparison.png
