+++ experiments/parameter_exploration/results/mu_comparison.png
