+++ src/model/growth_cone.py
@@ -3,8 +3,6 @@ Module providing Growth Cone class for growth cone representation.
 """
 import math
 
-from model.growth_cone.history import History
-
 
 class GrowthCone:
     """
@@ -47,8 +45,6 @@ class GrowthCone:
 
         self.history = History(self.potential, self.adap_co, self.pos_current, self.ligand_start, self.receptor_start,
                                self.reset_force_receptor, self.reset_force_ligand)
-
-        self.mark = False
 
     def __str__(self):
         """
@@ -116,3 +112,36 @@ class GrowthCone:
         """
         self.receptor_current += knock_in
         self.ligand_current = 0.35 / self.receptor_current
+
+
+class History:
+    def __init__(self, potential_ini, adap_co_ini, position_ini, ligand_ini, receptor_ini,
+                 reset_force_receptor_ini, reset_force_ligand_ini):
+        self.potential = [potential_ini]
+        self.adap_co = [adap_co_ini]
+        self.position = [position_ini]
+        self.ligand = [ligand_ini]
+        self.receptor = [receptor_ini]
+        self.reset_force_receptor = [reset_force_receptor_ini]
+        self.reset_force_ligand = [reset_force_ligand_ini]
+
+    def update_potential(self, potential_new):
+        self.potential.append(potential_new)
+
+    def update_adap_co(self, adap_co_new):
+        self.adap_co.append(adap_co_new)
+
+    def update_position(self, adap_position_new):
+        self.position.append(adap_position_new)
+
+    def update_ligand(self, ligand_new):
+        self.ligand.append(ligand_new)
+
+    def update_receptor(self, receptor_new):
+        self.receptor.append(receptor_new)
+
+    def update_reset_force_receptor(self, reset_force_receptor_new):
+        self.reset_force_receptor.append(reset_force_receptor_new)
+
+    def update_reset_force_ligand(self, reset_force_ligand_new):
+        self.reset_force_ligand.append(reset_force_ligand_new)
